+++ Changelog.md
@@ -12,6 +12,7 @@ Showroom View: Component presentation view for showcasing the library on a devic
 - MessageView  Component: Added a reusable component featuring an icon, a primary title, and a secondary message to display status or greetings.
 - Logos and iso added to the brand components
 - Added TopBars
+- Squared and Rectangle button component added.
 - added Toast Component
 
 ## [0.0.1] - 2025-10-7